split color-coded-util into constants and functions files
