+++ changelog.txt
@@ -2,6 +2,7 @@
 Version: 1.4.0
 Date: ????
   Changes:
+    - split color-coded-util file into constants and functions files
 ---------------------------------------------------------------------------------------------------
 Version: 1.3.1
 Date: 2025-06-11
